Coffee Machine project
works as expected, notes to consider moving forward.
Build out dictionaries better, e.g. create an ingrediant dict within the drink dict, this would allow me to dynamically call which dict for use as function para and would allow me to iterate through the dict in the function removing the need for if statements to determine what drink was choosen.